+++ src/main/java/com/example/instaclone/domian/likes/LikesRepository.java
@@ -4,12 +4,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 
-public interface LikesRepository extends JpaRepository<Likes,Integer> {
+public interface LikesRepository extends JpaRepository<Likes,Long> {
     @Modifying
     @Query(value = "INSERT INTO likes(imageId,userId,createDate) VALUES(:imageId, :principalId,now())",nativeQuery = true)
-    int mLikes(int imageId,int principalId);
+    int mLikes(Long imageId,Long principalId);
 
     @Modifying
     @Query(value = "DELETE FROM likes WHERE imageId =:imageId AND userId = :principalId",nativeQuery = true)
-    int mUnlikes(int imageId,int principalId);
+    int mUnlikes(Long imageId,Long principalId);
 }
